VBoxFBOverlay.cpp: fixed signed / unsigned warning caused by the framebuffer width() and height() methods returning ulong. The really confusing part is how they store the value as signed int internally...
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@24668 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/VBoxFBOverlay.cpp
@@ -5251,8 +5251,8 @@ bool VBoxQGLOverlay::vboxSynchGl()
             && mContainer->address() == mpOverlayWidget->vboxAddress()
             && mContainer->bitsPerPixel() == mpOverlayWidget->vboxBitsPerPixel()
             && mContainer->bytesPerLine() == mpOverlayWidget->vboxBytesPerLine()
-            && mContainer->width() == mpOverlayWidget->vboxFbWidth()
-            && mContainer->height() == mpOverlayWidget->vboxFbHeight())
+            && (int)mContainer->width() == mpOverlayWidget->vboxFbWidth()
+            && (int)mContainer->height() == mpOverlayWidget->vboxFbHeight())
     {
         return false;
     }
